Thread-safe determination of TObject::IsOnHeap()
The previous mechanism of determining TObject::IsOnHeap() was found
to be inaccurate and thread unsafe. The new way is to fill the memory
area returned by calls to TObject::new with a special value and
as the first operation in TObject's constructor look to see if the
uninitialized member data have that value set. If they do we assume
the object was made via a call to 'new' and therefore is on the heap.

The interface in TStorage used to implement the old mechanism has
been declared obsolete via the use of ::Obsolete.