Implement TH1::Clone
The previous Clone method called for a histogram relied on both
the I/O system and on Cint to function. This lead to either
data race conditions or synchronization points when working with
different  histograms on different threads. Using TClass to
create an empty instance and then calling Copy plus a bit of
cleanup allows different histograms to be uncoupled.